toy/requirements.txt
absl-py==0.14.0
aiohttp==3.7.4.post0
alembic==1.7.5
antlr4-python3-runtime==4.8
anyio==3.4.0
argcomplete==1.12.3
argon2-cffi-bindings==21.2.0
argon2-cffi==21.2.0
async-timeout==3.0.1
attrs==20.3.0
autopage==0.4.0
autopep8==1.5.7
babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.1
black==21.12b0
bleach==4.1.0
blessings==1.7
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.3
cliff==3.10.0
cmaes==0.8.2
cmd2==2.3.3
colorama==0.4.4
colorlog==6.6.0
commonmark==0.9.1
configparser==5.1.0
cycler==0.10.0
cython==0.29.24
debugpy==1.4.1
decorator==5.1.0
defusedxml==0.7.1
dill==0.3.3
distlib==0.3.4
docker-pycreds==0.4.0
einops==0.3.2
entrypoints==0.3
filelock==3.4.0
flake8==4.0.1
fsspec==2021.9.0
future==0.18.2
gitdb==4.0.9
gitpython==3.1.24
google-auth-oauthlib==0.4.6
google-auth==1.35.0
gpustat==0.6.0
gputil==1.4.0
greenlet==1.1.2
grpcio==1.41.0
hydra-colorlog==1.1.0
hydra-core==1.1.0
hydra-joblib-launcher==1.1.5
identify==2.4.0
idna==3.2
importlib-metadata==4.8.1
importlib-resources==5.4.0
iniconfig==1.1.1
ipdb==0.13.9
ipykernel==6.2.0
ipython-genutils==0.2.0
ipython==7.28.0
isort==5.10.1
jammy==0.0.5
jedi==0.18.0
jinja2==3.0.3
joblib==1.0.1
json5==0.9.6
jsonschema==4.2.1
jupyter-client==7.0.1
jupyter-core==4.7.1
jupyter-server==1.13.1
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab==3.2.5
kiwisolver==1.3.1
line-profiler==3.4.0
llvmlite==0.37.0
lmdb==1.2.1
loguru==0.5.3
mako==1.1.6
markdown==3.3.4
markupsafe==2.0.1
matplotlib-inline==0.1.3
matplotlib==3.4.2
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
msgpack-numpy==0.4.7.1
msgpack==1.0.3
multidict==5.1.0
multipledispatch==0.6.0
mypy-extensions==0.4.3
nbclassic==0.3.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
nodeenv==1.6.0
notebook==6.4.6
numba==0.54.1
numpy==1.21.5
nvidia-ml-py3==7.352.0
oauthlib==3.1.1
omegaconf==2.1.1
optuna==2.4.0
ordered-set==4.0.2
packaging==21.0
pandas==1.3.0
pandocfilters==1.5.0
parse==1.19.0
parso==0.8.2
pathspec==0.9.0
pathtools==0.1.2
pbr==5.8.0
pexpect==4.8.0
pickleshare==0.7.5
pillow==8.3.1
pip==21.1.3
platformdirs==2.4.0
pluggy==1.0.0
pre-commit==2.16.0
prettytable==2.4.0
prometheus-client==0.12.0
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.18.0
psutil==5.8.0
ptyprocess==0.7.0
pudb==2021.2.2
py==1.11.0
pyarrow==6.0.1
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydeprecate==0.3.1
pyflakes==2.4.0
pygments==2.10.0
pyparsing==2.4.7
pyperclip==1.8.2
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
python-memcached==1.59
pytorch-lightning==1.5.5
pytz==2021.1
pyyaml==5.4.1
pyzmq==22.3.0
requests-oauthlib==1.3.0
requests==2.26.0
rich==10.15.2
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.0
sdeint==0.2.1
sdepy==1.1.2
seaborn==0.11.1
send2trash==1.8.0
sentry-sdk==1.4.3
setuptools==58.0.4
sh==1.14.2
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
sqlalchemy==1.4.28
stevedore==3.5.0
subprocess32==3.5.4
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboard==2.6.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==2.2.0
toml==0.10.2
tomli==1.2.2
torch-fidelity==0.3.0
torch==1.10.1
torchaudio==0.10.1
torchinfo==1.5.4
torchmetrics==0.6.1
torchvision==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
typed-ast==1.5.1
typing-extensions==3.10.0.2
ubelt==0.10.2
urllib3==1.26.7
urwid-readline==0.13
urwid==2.1.2
virtualenv==20.10.0
wandb==0.10.33
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
werkzeug==2.0.1
wheel==0.37.0
yarl==1.6.3
yaspin==2.1.0
zipp==3.6.0